% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\docType{class}
\name{occurrence}
\alias{occurrence}
\alias{occurrence-class}
\title{EarthLife API occurrence objects.}
\description{
This class contains the results from calls to the EarthLife Consortium, generated by calls to
\code{get_by_taxon} and others.
}
\section{Slots}{

   \describe{
     \item{\code{records}:}{Object of class \code{"data.frame"}.}
     \item{\code{meta}:}{Object of class \code{"list"} containing metadata about the call, date of record and data source.}
   }
}
\author{
Simon Goring
}

