% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrence.R
\name{get_occurrence}
\alias{get_occurrence}
\title{Return occurrences for taxa within the PBDB/Neotoma.}
\usage{
get_occurrence(x, occ_id, taxon_name, base_name, match_name, base_id, taxon_id,
  site_id, bbox, vocab)
}
\arguments{
\item{x}{An optional value, either a \code{numeric} site ID or object of class \code{download}, \code{download_list} or \code{site}.}

\item{occ_id}{A numeric Occurrence ID.}

\item{taxon_name}{The taxon name as a \code{character}}

\item{base_name}{The taxon name as a \code{character}}

\item{match_name}{The taxon name as a \code{character}}

\item{base_id}{The taxon name as a \code{character}}

\item{taxon_id}{The taxon name as a \code{character}}

\item{site_id}{The taxon name as a \code{character}}

\item{bbox}{The taxon name as a \code{character}}

\item{vocab}{The taxonomic vocabulary to use, either \'pbdb\' or \'neotoma\'.}
}
\value{
More details on the use of these parameters can be obtained from
   \url{https://training.paleobiodb.org/comp1.0/}.

   A list of class `data.frame`.  , with each item corresponding to an individual record.
   Searches that return no items will result in a NULL value being returned.
   Otherwise each list item (each dataset record) includes the following components:

 \item{ \code{dataset.id} }{Unique database record identifier for the dataset.}
 \item{ \code{dataset.name}  }{Name of the dataset; not commonly used.}
 \item{ \code{CollUnitHandle}  }{Code name of the Collection Unit with which the dataset is associated. This code may be up to 10 characters. Data are frequently distributed by Collection Unit, and the Handle is used for file names.}
 \item{ \code{CollUnitID}  }{Unique database record identifier for the collection unit.}
 \item{ \code{CollType}  }{The collection type. Types include cores, sections, excavations, and animal middens.}
 \item{ \code{DatasetType}  }{The dataset type, such as: geochronologic, loss-on-ignition, pollen, plant macrofossils, vertebrate fauna, etc.}
 \item{ \code{AgeOldest}  }{The oldest of all sample ages (in calendar years before present) in the dataset.}
 \item{ \code{AgeYoungest}  }{The youngest of all sample ages (in calendar years before present) in the dataset.}
 \item{ \code{SubDates}  }{An array of objects that describe dataset submission events.  If multiple submissions occurred then this is a table.}
 \item{ \code{DatasetPIs}  }{An array of objects that describe Principal Investigators associated with a dataset.}
 \item{ \code{Site}  }{An object describing the site where the dataset samples were taken.}
}
\description{
A wrapper for the Composite API, returning all records from both datasets.
}
\examples{
\dontrun{
# Search for sites with "Thuja" pollen that are older than 8kyr BP and
# that are on the west coast of North America:
t8kyr.datasets <- get_dataset(taxonname='Thuja*', loc=c(-150, 20, -100, 60), ageyoung = 8000)

# Search for vertebrate fossils in Canada (gpid: 756) within the last 2kyr.
gpids <- get_table(table.name='GeoPoliticalUnits')
canID <- gpids[which(gpids$GeoPoliticalName == 'Canada'),1]

v2kyr.datasets <- get_dataset(datasettype='vertebrate fauna', gpid=canID, ageold = 2000)
}
}
\author{
Simon J. Goring \email{goring@wisc.edu}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  https://training.paleobiodb.org/comp1.0
}
\keyword{IO}
\keyword{connection}

